+++ src/PlutoMathInput.jl
@@ -5,13 +5,14 @@ using AbstractPlutoDingetjes.Bonds
 using HypertextLiteral
 using JSON3
 
-export MathInput
+export MathInput, MathDisplay
 
 # Stubs for Symbolics extension
 function to_symbolics end
 function from_symbolics end
 
 include("mathinput.jl")
+include("mathdisplay.jl")
 include("show.jl")
 
 end # module